refactor: extract the racing class from the function
